+++ app/services/create_fund_activity.rb
@@ -8,7 +8,8 @@ class CreateFundActivity
   def call
     activity = Activity.new
     activity.organisation = Organisation.find(organisation_id)
-    activity.reporting_organisation_reference = activity.organisation.iati_reference
+    activity.reporting_organisation = activity.organisation
+    activity.extending_organisation = service_owner
 
     activity.wizard_status = "blank"
     activity.level = :fund
@@ -21,9 +22,13 @@ class CreateFundActivity
     activity.accountable_organisation_reference = "GB-GOV-13"
     activity.accountable_organisation_type = "10"
 
-    activity.extending_organisation = Organisation.find_by!(service_owner: true)
-
     activity.save(validate: false)
     activity
+  end
+
+  private
+
+  def service_owner
+    Organisation.find_by_service_owner(true)
   end
 end